refactor(album): migrate constants and types to albumConfig for better organization
